genAITemplate/src/apis/chat/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework==1.0.0rc2",
    "agent-framework-foundry-local>=1.0.0b260225",
    "fastapi>=0.133.1",
    "fastapi-cli>=0.0.24",
    "foundry-local-sdk>=0.5.1",
    "python-dotenv>=1.2.1",
]
